CI: use circleci (#76)
- Replace travis with circleci
- Update the drivers to its latest versions
- Update selenium pytest
- Test with more versions of sphinx and python
- Inject the extra js data before the head tag, as latest versions of sphinx don't inject jquery with the tag that the test was using.
- Fix tests for firefox (this test has always been failing for me locally before moving to circle)

The problem with the test was that it was assuming that the inner real width/height was the same the given by the test, but this doesn't count with extra space generated for window borders, navigation bar, etc (`set_viewport_size` uses this too).
This was failing only for firefox. Also, the current test was just assuming the margins were the same size (dividing by 2), instead I'm just checking that the margins are almost the same size (since we want to check that the modal is in the middle).